Move Constants interface to org.saidone.misc package and update import in BaseComponent accordingly
