+++ src/main/java/org/saidone/component/BaseComponent.java
@@ -21,7 +21,7 @@ package org.saidone.component;
 import jakarta.annotation.PostConstruct;
 import jakarta.annotation.PreDestroy;
 import lombok.extern.slf4j.Slf4j;
-import org.saidone.Constants;
+import org.saidone.misc.Constants;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.ConfigurableApplicationContext;
